core.controller: try to fix depth-related issues in website_export
This is not the right solution. We should do a breadth-first search,
in order to accurately honour the depth parameter. But then, we would
have to make a 2-pass process: first, generate the documents and build
the tree and url conversion; second, fix hyperlinks in generated
documents. To be done...

git-svn-id: svn+ssh://svn.gna.org/svn/advene/trunk@5791 3fb6f576-06d4-0310-ad11-a4b695aa9f05