gui.views.timeline: optimize the timeline loading by resizing the layout widget.
It looks like the cost of the timeline update mainly lies in the
handling of expose events as annotation widgets are created. If we
simply do self.layout.hide() then the .show() call at the end takes a
while. If we keep the self.layout shown but resize its container so
that it is not visible, we prevent expose events to be handled, and
the loading is smoother.

git-svn-id: svn+ssh://svn.gna.org/svn/advene/trunk@7478 3fb6f576-06d4-0310-ad11-a4b695aa9f05