gui.views.timeline: fix concurrency issues in update_model calls
update_model could be called either from the __init__ method (in the
expose-event handler), or from the PackageActivate event handler, or
even from the DurationUpdate event handler. Depending on which comes
first, there is a possibility that order vary from one invocation to
another.

Moreover, since we switched to asynchronous populate() method, we have
to add a lock to avoid multiple concurrent calls to the method.

git-svn-id: svn+ssh://svn.gna.org/svn/advene/trunk@7497 3fb6f576-06d4-0310-ad11-a4b695aa9f05