gui.edit.importer: allow non-local filenames (URIs for instance)
Some filters are able to process URIs, offer the possibility to enter
them.
The gtk.FileChooser offers a set_local(False) feature to allow
non-local filenames, but gtk.FileChooserButton does not reflect the
selected URI if it is not a valid local filename. Thus we implemented
a workaround to explicitly update the Button label information.

git-svn-id: svn+ssh://svn.gna.org/svn/advene/trunk@7508 3fb6f576-06d4-0310-ad11-a4b695aa9f05