gui.plugins.shotvalidation: have shotvalidation take focus
Prev/next shot navigation shorcuts were already implemented (using Page Up/Page Down keys) but since the view is often launched from the timeline, which tries to get focus on mouse over, the focus remained on timeline instead of the newly opened view.
A solution is to click on the "Current time" button.
This patch tries to grab the focus after some timeout, in order to have the user move the cursor out of the timeline view.

Closes #31